requirements.txt
Django~=2.1.7
djangorestframework==3.9.2
django-filter==2.1.0
Faker==1.0.2
factory-boy==2.11.1
text-unidecode==1.2
django-extensions==2.1.6